+++ src/look_ahead.rs
@@ -110,15 +110,6 @@ fn filter<'a>(
     context: &'a Context<'a>,
 ) {
     for item in &selection_set.items {
-        // doing this imperatively is a bit nasty, but using iterators would
-        // require a boxed return type (I believe) as its recusive
-
-        // ignore any items that are skipped (i.e. @skip/@include)
-        if context.is_skip(&item.node.directives()).unwrap_or(false) {
-            // TODO: should we throw errors here? they will be caught later in execution and it'd cause major backwards compatibility issues
-            continue;
-        }
-
         match &item.node {
             Selection::Field(field) => {
                 if field.node.name.node == name {