+++ src/validation/rules/no_unused_variables.rs
@@ -1,12 +1,13 @@
 use std::collections::{HashMap, HashSet};
 
+use async_graphql_value::Value;
+
 use crate::parser::types::{
     ExecutableDocument, FragmentDefinition, FragmentSpread, OperationDefinition, VariableDefinition,
 };
 use crate::validation::utils::{referenced_variables, Scope};
 use crate::validation::visitor::{Visitor, VisitorContext};
 use crate::{Name, Pos, Positioned};
-use async_graphql_value::Value;
 
 #[derive(Default)]
 pub struct NoUnusedVariables<'a> {