+++ tests/directive.rs
@@ -12,20 +12,37 @@ pub async fn test_directive_skip() {
     }
 
     let schema = Schema::new(QueryRoot, EmptyMutation, EmptySubscription);
-    let resp = schema
+    let data = schema
         .execute(
             r#"
-            {
+            fragment A on QueryRoot {
+                value5: value @skip(if: true)
+                value6: value @skip(if: false)
+            }
+            
+            query {
                 value1: value @skip(if: true)
                 value2: value @skip(if: false)
+                ... @skip(if: true) {
+                    value3: value
+                }
+                ... @skip(if: false) {
+                    value4: value
+                }
+                ... A
             }
         "#,
         )
-        .await;
+        .await
+        .into_result()
+        .unwrap()
+        .data;
     assert_eq!(
-        resp.data,
+        data,
         value!({
             "value2": 10,
+            "value4": 10,
+            "value6": 10,
         })
     );
 }
@@ -56,62 +73,6 @@ pub async fn test_directive_include() {
         resp.data,
         value!({
             "value1": 10,
-        })
-    );
-}
-
-#[tokio::test]
-pub async fn test_directive_ifdef() {
-    struct QueryRoot;
-
-    #[Object]
-    impl QueryRoot {
-        pub async fn value1(&self) -> i32 {
-            10
-        }
-    }
-
-    struct MutationRoot;
-
-    #[Object]
-    impl MutationRoot {
-        pub async fn action1(&self) -> i32 {
-            10
-        }
-    }
-
-    let schema = Schema::new(QueryRoot, MutationRoot, EmptySubscription);
-    let resp = schema
-        .execute(
-            r#"
-            {
-                value1 @ifdef
-                value2 @ifdef
-            }
-        "#,
-        )
-        .await;
-    assert_eq!(
-        resp.data,
-        value!({
-            "value1": 10,
-        })
-    );
-
-    let resp = schema
-        .execute(
-            r#"
-            mutation {
-                action1 @ifdef
-                action2 @ifdef
-            }
-        "#,
-        )
-        .await;
-    assert_eq!(
-        resp.data,
-        value!({
-            "action1": 10,
         })
     );
 }