+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.0.5] 2021-11-19
+
+- Remove skipped fields from the document before executing the query.
+- Add `isRepeatable` field to `__Directive` - [GraphQL - October 2021]
+
 ## [3.0.4] 2021-11-18
 
 - Remove `OutputJson` because `Json` can replace it.